4.5.0

+++ src/module/chat/chat-listeners.js
@@ -5,6 +5,7 @@
 
 import {buildActorInfo, formatDice, getChatBaseData} from './base-messages'
 import {TokenManager} from '../pixi/token-manager'
+import {DLAfflictions} from '../active-effects/afflictions'
 import { changesMatch } from '../utils/chat'
 
 const tokenManager = new TokenManager()
@@ -57,9 +58,13 @@ async function _onChatRollDamage(event) {
   const token = li.closest('.demonlord')
   const actor = _getChatCardActor(token)
 
-  for (let effect of tokenManager.getTokenByActorId(actor.id).actor.appliedEffects) {
-    const specialDuration = foundry.utils.getProperty(effect, 'flags.specialDuration')
-    if (specialDuration === 'NextDamageRoll') await effect?.delete()
+  const appliedEffects = tokenManager.getTokenByActorId(actor.id)?.actor?.appliedEffects
+  
+  if (appliedEffects?.length) {
+    for (let effect of appliedEffects) {
+      const specialDuration = foundry.utils.getProperty(effect, 'flags.specialDuration')
+      if (specialDuration === 'NextDamageRoll') await effect?.delete()
+    }
   }
 
   const item = li.children[0]
@@ -353,7 +358,9 @@ async function _onChatMakeChallengeRoll(event) {
   const start = li.closest('.demonlord')
   const boonsbanesEntered = start.children[1].children[0].children[0].children[1]?.value
 
-  await actor.rollAttribute(attribute, parseInt(boonsbanes) + parseInt(boonsbanesEntered), 0)
+  if (!DLAfflictions.isActorBlocked(actor, 'challenge', attributeName)) {
+    await actor.rollAttribute(attribute, parseInt(boonsbanes) + parseInt(boonsbanesEntered), 0)
+  }
 }
 
 /* -------------------------------------------- */
